refactor: MediaContent 내부에서 ContentList 분리
